Restrukturer steg-tester
