+++ behandlingsprosess/src/test/java/no/nav/foreldrepenger/behandling/steg/uttak/fp/KontrollerfaktaUttakStegTest.java
@@ -1,6 +1,5 @@
 package no.nav.foreldrepenger.behandling.steg.uttak.fp;
 
-import static no.nav.foreldrepenger.behandlingslager.behandling.aksjonspunkt.AksjonspunktDefinisjon.MANUELL_KONTROLL_AV_OM_BRUKER_HAR_ALENEOMSORG;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDate;
@@ -17,9 +16,7 @@ import no.nav.foreldrepenger.behandlingskontroll.BehandlingskontrollKontekst;
 import no.nav.foreldrepenger.behandlingskontroll.FagsakYtelseTypeRef;
 import no.nav.foreldrepenger.behandlingskontroll.transisjoner.FellesTransisjoner;
 import no.nav.foreldrepenger.behandlingslager.behandling.Behandling;
-import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingType;
 import no.nav.foreldrepenger.behandlingslager.behandling.aksjonspunkt.AksjonspunktDefinisjon;
-import no.nav.foreldrepenger.behandlingslager.behandling.personopplysning.RelasjonsRolleType;
 import no.nav.foreldrepenger.behandlingslager.behandling.personopplysning.SivilstandType;
 import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepository;
 import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepositoryProvider;
@@ -30,8 +27,6 @@ import no.nav.foreldrepenger.behandlingslager.behandling.ytelsefordeling.periode
 import no.nav.foreldrepenger.behandlingslager.fagsak.FagsakYtelseType;
 import no.nav.foreldrepenger.behandlingslager.geografisk.Landkoder;
 import no.nav.foreldrepenger.behandlingslager.testutilities.behandling.ScenarioFarSøkerForeldrepenger;
-import no.nav.foreldrepenger.behandlingslager.testutilities.behandling.ScenarioMorSøkerForeldrepenger;
-import no.nav.foreldrepenger.behandlingslager.testutilities.behandling.personopplysning.PersonAdresse;
 import no.nav.foreldrepenger.dbstoette.CdiDbAwareTest;
 import no.nav.foreldrepenger.domene.typer.AktørId;
 import no.nav.foreldrepenger.domene.uttak.fakta.KontrollerFaktaUttakTjeneste;
@@ -102,70 +97,5 @@ public class KontrollerfaktaUttakStegTest {
         assertThat(aksjonspunkter).hasSize(1);
         assertThat(aksjonspunkter.get(0)).isEqualTo(AksjonspunktDefinisjon.MANUELL_KONTROLL_AV_OM_BRUKER_HAR_OMSORG);
 
-    }
-
-    @Test
-    public void skal_utledede_aksjonspunkt_basert_på_fakta_om_foreldrepenger_til_mor() {
-        // oppsett
-        byggBehandingMedMorSøkerTypeOgHarAleneOmsorgOgAnnenforelderHarIkkeRett();
-
-        var input = uttakInputTjeneste.lagInput(behandling);
-        var resultat = tjeneste.utledAksjonspunkter(input);
-
-        // Assert
-        assertThat(resultat).containsExactlyInAnyOrder(MANUELL_KONTROLL_AV_OM_BRUKER_HAR_ALENEOMSORG);
-
-    }
-
-    private Behandling byggBehandingMedMorSøkerTypeOgHarAleneOmsorgOgAnnenforelderHarIkkeRett() {
-        var AKTØR_ID_MOR = AktørId.dummy();
-        var AKTØR_ID_FAR = AktørId.dummy();
-
-        var scenario = ScenarioMorSøkerForeldrepenger.forFødselMedGittAktørId(AKTØR_ID_MOR);
-        var rettighet = new OppgittRettighetEntitet(false, true, true);
-
-        var builderForRegisteropplysninger = scenario.opprettBuilderForRegisteropplysninger();
-
-        var bostedsadresse = PersonAdresse.builder().adresselinje1("Portveien 2").postnummer("7000").land(Landkoder.NOR);
-
-        var annenPrt = builderForRegisteropplysninger
-                .medPersonas()
-                .mann(AKTØR_ID_FAR, SivilstandType.GIFT)
-                .bostedsadresse(bostedsadresse)
-                .relasjonTil(AKTØR_ID_MOR, RelasjonsRolleType.EKTE, true)
-                .build();
-        scenario.medRegisterOpplysninger(annenPrt);
-
-        var søker = builderForRegisteropplysninger
-                .medPersonas()
-                .kvinne(AKTØR_ID_MOR, SivilstandType.GIFT)
-                .bostedsadresse(bostedsadresse)
-                .statsborgerskap(Landkoder.NOR)
-                .relasjonTil(AKTØR_ID_FAR, RelasjonsRolleType.EKTE, true)
-                .build();
-
-        scenario.medRegisterOpplysninger(søker);
-
-        scenario.medSøknadAnnenPart()
-                .medAktørId(AKTØR_ID_FAR);
-
-        scenario.medSøknad();
-        scenario.medOppgittRettighet(rettighet);
-        var hendelseBuilder = scenario.medSøknadHendelse();
-        var termindato = LocalDate.now().plusDays(35);
-        hendelseBuilder.medTerminbekreftelse(hendelseBuilder.getTerminbekreftelseBuilder()
-                .medNavnPå("asdf")
-                .medUtstedtDato(LocalDate.now())
-                .medTermindato(termindato));
-
-        scenario.medFordeling(new OppgittFordelingEntitet(Collections.singletonList(
-                OppgittPeriodeBuilder.ny()
-                        .medPeriode(termindato, termindato.plusWeeks(6))
-                        .medPeriodeType(UttakPeriodeType.FORELDREPENGER)
-                        .build()),
-                true));
-
-        behandling = scenario.lagre(repositoryProvider);
-        return behandling;
     }
 }
